use `number` and getters in npm public API (#463)
Turns out that using `#[napi(ts_return_type = "number")]` only affects
the return type in the generated `index.d.ts` file, but not the actual
return type at runtime. Currently we are using `usize`, which maps to
JavaScript's `BigInt` by default, which means our types are incorrect.

This PR uses `u32` instead which maps to JavaScript's `number` type.
Also adds a few missing `getter` attributes, and adds a test to check
the new types.

Follow up on #458

+++ crates/codegen/syntax_templates/src/typescript/cst_types.rs
@@ -34,30 +34,30 @@ impl RuleNode {
     }
 
     #[napi(getter, ts_return_type = "[ start: number, end: number ]")]
-    pub fn byte_range(&self) -> [usize; 2] {
+    pub fn byte_range(&self) -> [u32; 2] {
         let range = self.0.range();
-        [range.start.byte, range.end.byte]
+        [range.start.byte as u32, range.end.byte as u32]
     }
 
     #[napi(getter, ts_return_type = "[ start: number, end: number ]")]
-    pub fn char_range(&self) -> [usize; 2] {
+    pub fn char_range(&self) -> [u32; 2] {
         let range = self.0.range();
-        [range.start.char, range.end.char]
+        [range.start.char as u32, range.end.char as u32]
     }
 
     #[napi(getter, ts_return_type = "[ start: number, end: number ]")]
-    pub fn byte_range_including_trivia(&self) -> [usize; 2] {
+    pub fn byte_range_including_trivia(&self) -> [u32; 2] {
         let range = self.0.range_including_trivia();
-        [range.start.byte, range.end.byte]
+        [range.start.byte as u32, range.end.byte as u32]
     }
 
     #[napi(getter, ts_return_type = "[ start: number, end: number ]")]
-    pub fn char_range_including_trivia(&self) -> [usize; 2] {
+    pub fn char_range_including_trivia(&self) -> [u32; 2] {
         let range = self.0.range_including_trivia();
-        [range.start.char, range.end.char]
+        [range.start.char as u32, range.end.char as u32]
     }
 
-    #[napi(ts_return_type = "(RuleNode | TokenNode)[]")]
+    #[napi(getter, ts_return_type = "(RuleNode | TokenNode)[]")]
     pub fn children(&self, env: Env) -> Vec<JsObject> {
         match self.0.as_ref() {
             Node::Rule { children, .. } => children.iter().map(|child| child.to_js(&env)).collect(),
@@ -82,30 +82,30 @@ impl TokenNode {
     }
 
     #[napi(getter, ts_return_type = "[ start: number, end: number ]")]
-    pub fn byte_range(&self) -> [usize; 2] {
+    pub fn byte_range(&self) -> [u32; 2] {
         let range = self.0.range();
-        [range.start.byte, range.end.byte]
+        [range.start.byte as u32, range.end.byte as u32]
     }
 
     #[napi(getter, ts_return_type = "[ start: number, end: number ]")]
-    pub fn char_range(&self) -> [usize; 2] {
+    pub fn char_range(&self) -> [u32; 2] {
         let range = self.0.range();
-        [range.start.char, range.end.char]
+        [range.start.char as u32, range.end.char as u32]
     }
 
     #[napi(getter, ts_return_type = "[ start: number, end: number ]")]
-    pub fn byte_range_including_trivia(&self) -> [usize; 2] {
+    pub fn byte_range_including_trivia(&self) -> [u32; 2] {
         let range = self.0.range_including_trivia();
-        [range.start.byte, range.end.byte]
+        [range.start.byte as u32, range.end.byte as u32]
     }
 
     #[napi(getter, ts_return_type = "[ start: number, end: number ]")]
-    pub fn char_range_including_trivia(&self) -> [usize; 2] {
+    pub fn char_range_including_trivia(&self) -> [u32; 2] {
         let range = self.0.range_including_trivia();
-        [range.start.char, range.end.char]
+        [range.start.char as u32, range.end.char as u32]
     }
 
-    #[napi(ts_return_type = "(RuleNode | TokenNode)[]")]
+    #[napi(getter, ts_return_type = "(RuleNode | TokenNode)[]")]
     pub fn trivia(&self, env: Env) -> Vec<JsObject> {
         match self.0.as_ref() {
             Node::Token { trivia, .. } => {
